requirements.txt
pyyaml
requests
jinja2
pydantic>=2.0
litellm>=1.75.5
tenacity
rich
python-dotenv
typer
openpyxl
matplotlib
numpy
platformdirs
ghapi
fastcore
